requirements.txt
fastapi
uvicorn
pillow
imagehash
opencv-python-headless
python-multipart
google-generativeai
python-dotenv
